accounts/views/profile_views.py
from rest_framework import viewsets
from accounts.models import Profile
from accounts.serializers import ProfileSerializer

class ProfileViewSet(viewsets.ModelViewSet):
    queryset = Profile.objects.all()
    serializer_class = ProfileSerializer