requirements.txt
Django
psycopg2
djangorestframework
drf-spectacular
pillow